Update Wallet API
Fix amount bugs.

+++ assets/api/wallet/delete.php
@@ -1,6 +1,6 @@
 <?php
     include_once( realpath( dirname( __FILE__ ) . "/../../config/config.php" ) );
-    include_once( MODULES_PATH . "/wallet.php" );
+    include_once( MODULES_PATH . "wallet.php" );
 
     if ( isset( $_POST ) )
     {
@@ -10,6 +10,11 @@
         $wallet_data_connector = new Wallet_Data_Connector();
         $wallet = $wallet_data_connector->read( $conn, $wallet );
         $wallet = $wallet_data_connector->convert( $wallet );
+        if ( $wallet->get( 'status' ) == "Default" )
+        {
+            echo json_encode( array( "result" => false, "message" => "This wallet cannot delete" ) );
+            die();
+        }
         $res = $wallet_data_connector->delete( $conn, $wallet );
 
         if ( $res )
